requirements.txt
absl-py==0.9.0
aniso8601==8.0.0
astor==0.8.1
Babel==2.8.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
constants==0.6.0
cycler==0.10.0
Flask==1.1.1
Flask-Babel==0.12.2
Flask-MySQL==1.5.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
Flask-Table==0.5.0
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mysqlclient==1.4.6
numexpr==2.7.1
numpy==1.18.0
oauthlib==3.1.0
opencv-python-headless==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
pconst==1.0.0
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
six==1.13.0
SQLAlchemy==1.3.12
tables==3.6.1
tb-nightly==2.2.0a20191227
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tf-estimator-nightly==2.0.0.dev2019122709
tf-nightly==2.1.0.dev20191227
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
